requirements.txt
Django==4.2.11
Pillow==10.1.0
djangorestframework==3.15.1
django-filter==24.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
dj-rest-auth==5.1.0
cloudinary==1.39.1
dj3-cloudinary-storage
django-cloudinary-storage==0.3.0
django-hitcount
django-environ==0.11.2
django-import-export==3.3.7
python-dotenv==1.0.1
pypdf==4.2.0
google-generativeai==0.4.1